Updated choices generation
World file now has 3 json lists for storing the json of locations, characters and items for use in the choices flow on template. The template has been updated. Initialisation adds the jsons to world file too. In the game loop after we get our jsons they must be added to the world json lists.